Moved lib/api tests out of subfolder into main folder.
