+++ app/lib/api/ApiRequest.test.js
@@ -2,10 +2,10 @@
 
 import { API_URL } from 'config/api';
 
-import ApiRequest from '../ApiRequest';
-import { MEDIA_TYPE } from '../constants';
-import type { Request } from '../model';
-import { methodTypes } from '../model';
+import ApiRequest from './ApiRequest';
+import { MEDIA_TYPE } from './constants';
+import type { Request } from './model';
+import { methodTypes } from './model';
 
 const defaultHeaders = {
   'Content-Type': MEDIA_TYPE,
